+++ src/MiddlewareAPIs/EditorAPI.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+namespace OpenSim { class ComponentPath; }
 namespace OpenSim { class Coordinate; }
 namespace OpenSim { class Muscle; }
 namespace osc { class Popup; }
@@ -16,6 +17,7 @@ namespace osc
     public:
         virtual ~EditorAPI() noexcept = default;
 
+        virtual void pushComponentContextMenuPopup(OpenSim::ComponentPath const&) = 0;
         virtual void pushPopup(std::unique_ptr<Popup>) = 0;
         virtual void addMusclePlot(OpenSim::Coordinate const&, OpenSim::Muscle const&) = 0;
     };